+++ services/brig/src/Brig/API/Public.hs
@@ -92,6 +92,7 @@ import Servant.Swagger.UI
 import qualified System.Logger.Class as Log
 import Util.Logging (logFunction, logHandle, logTeam, logUser)
 import qualified Wire.API.Connection as Public
+import Wire.API.ErrorDescription
 import qualified Wire.API.Properties as Public
 import Wire.API.Routes.Public (EmptyResult (..))
 import qualified Wire.API.Routes.Public.Brig as BrigAPI
@@ -172,7 +173,7 @@ servantSitemap =
         BrigAPI.getUserQualified = getUserH,
         BrigAPI.getSelf = getSelf,
         BrigAPI.getHandleInfoUnqualified = getHandleInfoUnqualifiedH,
-        BrigAPI.getUserByHandleQualfied = getUserByHandleH,
+        BrigAPI.getUserByHandleQualified = getUserByHandleH,
         BrigAPI.listUsersByUnqualifiedIdsOrHandles = listUsersByUnqualifiedIdsOrHandles,
         BrigAPI.listUsersByIdsOrHandles = listUsersByIdsOrHandles,
         BrigAPI.getUserClientsUnqualified = getUserClientsUnqualified,
@@ -228,7 +229,7 @@ sitemap o = do
       Doc.description "Handle to check"
     Doc.response 200 "Handle is taken" Doc.end
     Doc.errorResponse invalidHandle
-    Doc.errorResponse handleNotFound
+    Doc.errorResponse (errorDescriptionToWai handleNotFound)
 
   -- some APIs moved to servant
   -- end User Handle API
@@ -385,7 +386,7 @@ sitemap o = do
     Doc.response 202 "Deletion is pending verification with a code." Doc.end
     Doc.response 200 "Deletion is initiated." Doc.end
     Doc.errorResponse badCredentials
-    Doc.errorResponse missingAuthError
+    Doc.errorResponse (errorDescriptionToWai missingAuthError)
 
   -- TODO put  where?
 
@@ -473,7 +474,7 @@ sitemap o = do
     Doc.response 204 "No change." Doc.end
     Doc.errorResponse connectionLimitReached
     Doc.errorResponse invalidTransition
-    Doc.errorResponse notConnected
+    Doc.errorResponse (errorDescriptionToWai notConnected)
     Doc.errorResponse invalidUser
 
   get "/connections/:id" (continue getConnectionH) $
@@ -771,7 +772,7 @@ getMultiUserPrekeyBundleUnqualifiedH :: UserId -> Public.UserClients -> Handler 
 getMultiUserPrekeyBundleUnqualifiedH zusr userClients = do
   maxSize <- fromIntegral . setMaxConvSize <$> view settings
   when (Map.size (Public.userClients userClients) > maxSize) $
-    throwStd tooManyClients
+    throwErrorDescription tooManyClients
   API.claimLocalMultiPrekeyBundles (ProtectedUser zusr) userClients !>> clientError
 
 getMultiUserPrekeyBundleH :: UserId -> Public.QualifiedUserClients -> Handler Public.QualifiedUserClientPrekeyMap
@@ -782,7 +783,7 @@ getMultiUserPrekeyBundleH zusr qualUserClients = do
           (\_ v -> Sum . Map.size $ v)
           (Public.qualifiedUserClients qualUserClients)
   when (size > maxSize) $
-    throwStd tooManyClients
+    throwErrorDescription tooManyClients
   API.claimMultiPrekeyBundles (ProtectedUser zusr) qualUserClients !>> clientError
 
 addClient :: UserId -> ConnId -> Maybe IpAddr -> Public.NewClient -> Handler BrigAPI.NewClientResponse
@@ -809,12 +810,8 @@ listClients :: UserId -> Handler [Public.Client]
 listClients zusr =
   lift $ API.lookupLocalClients zusr
 
-getClient :: UserId -> ClientId -> Handler (Union BrigAPI.GetClientResponse)
-getClient zusr clientId = do
-  mc <- lift $ API.lookupLocalClient zusr clientId
-  case mc of
-    Nothing -> throwEmptyForLegacyReasons status404
-    Just c -> Servant.respond (WithStatus @200 c)
+getClient :: UserId -> ClientId -> Handler (Maybe Public.Client)
+getClient zusr clientId = lift $ API.lookupLocalClient zusr clientId
 
 getUserClientsUnqualified :: UserId -> Handler [Public.PubClient]
 getUserClientsUnqualified uid = do
@@ -846,7 +843,7 @@ getUserClientQualified domain uid cid = do
 getClientCapabilities :: UserId -> ClientId -> Handler Public.ClientCapabilityList
 getClientCapabilities uid cid = do
   mclient <- lift (API.lookupLocalClient uid cid)
-  maybe (throwStd clientNotFound) (pure . Public.clientCapabilities) mclient
+  maybe (throwErrorDescription clientNotFound) (pure . Public.clientCapabilities) mclient
 
 getRichInfoH :: UserId ::: UserId ::: JSON -> Handler Response
 getRichInfoH (self ::: user ::: _) =
@@ -856,8 +853,12 @@ getRichInfo :: UserId -> UserId -> Handler Public.RichInfoAssocList
 getRichInfo self user = do
   -- Check that both users exist and the requesting user is allowed to see rich info of the
   -- other user
-  selfUser <- ifNothing userNotFound =<< lift (Data.lookupUser NoPendingInvitations self)
-  otherUser <- ifNothing userNotFound =<< lift (Data.lookupUser NoPendingInvitations user)
+  selfUser <-
+    ifNothing (errorDescriptionToWai userNotFound)
+      =<< lift (Data.lookupUser NoPendingInvitations self)
+  otherUser <-
+    ifNothing (errorDescriptionToWai userNotFound)
+      =<< lift (Data.lookupUser NoPendingInvitations user)
   case (Public.userTeam selfUser, Public.userTeam otherUser) of
     (Just t1, Just t2) | t1 == t2 -> pure ()
     _ -> throwStd insufficientTeamPermissions
@@ -944,16 +945,16 @@ createUser (Public.NewUserPublic new) = do
 
 getSelf :: UserId -> Handler Public.SelfProfile
 getSelf self =
-  lift (API.lookupSelfProfile self) >>= ifNothing userNotFound
+  lift (API.lookupSelfProfile self)
+    >>= ifNothing (errorDescriptionToWai userNotFound)
 
-getUserUnqualifiedH :: UserId -> UserId -> Handler Public.UserProfile
+getUserUnqualifiedH :: UserId -> UserId -> Handler (Maybe Public.UserProfile)
 getUserUnqualifiedH self uid = do
   domain <- viewFederationDomain
   getUserH self domain uid
 
-getUserH :: UserId -> Domain -> UserId -> Handler Public.UserProfile
-getUserH self domain uid =
-  ifNothing userNotFound =<< getUser self (Qualified uid domain)
+getUserH :: UserId -> Domain -> UserId -> Handler (Maybe Public.UserProfile)
+getUserH self domain uid = getUser self (Qualified uid domain)
 
 getUser :: UserId -> Qualified UserId -> Handler (Maybe Public.UserProfile)
 getUser self qualifiedUserId = API.lookupProfile self qualifiedUserId !>> fedError
@@ -1072,19 +1073,16 @@ checkHandlesH (_ ::: _ ::: req) = do
   return $ json (free :: [Handle])
 
 -- | This endpoint returns UserHandleInfo instead of UserProfile for backwards compatibility.
-getHandleInfoUnqualifiedH :: UserId -> Handle -> Handler Public.UserHandleInfo
+getHandleInfoUnqualifiedH :: UserId -> Handle -> Handler (Maybe Public.UserHandleInfo)
 getHandleInfoUnqualifiedH self handle = do
   domain <- viewFederationDomain
-  Public.UserHandleInfo . Public.profileQualifiedId <$> getUserByHandleH self domain handle
+  Public.UserHandleInfo . Public.profileQualifiedId
+    <$$> getUserByHandleH self domain handle
 
 -- | This endpoint returns UserProfile instead of UserHandleInfo to reduce
 -- traffic between backends in a federated scenario.
-getUserByHandleH :: UserId -> Domain -> Handle -> Handler Public.UserProfile
-getUserByHandleH self domain handle = do
-  maybeProfile <- Handle.getHandleInfo self (Qualified handle domain)
-  case maybeProfile of
-    Nothing -> throwStd handleNotFound
-    Just u -> pure u
+getUserByHandleH :: UserId -> Domain -> Handle -> Handler (Maybe Public.UserProfile)
+getUserByHandleH self domain handle = Handle.getHandleInfo self (Qualified handle domain)
 
 changeHandleH :: UserId ::: ConnId ::: JsonRequest Public.HandleUpdate -> Handler Response
 changeHandleH (u ::: conn ::: req) =